backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.32.0",
    "mcp>=1.13.1",
    "smolagents[litellm,mcp,toolkit]>=1.21.2",
]
